Corrected documentation in azure-authentication.md
Corrected the documentation on line 38-43, for mounting service principal authentication secret as a file. The way it is in the documentation at the moment did not work for me. After reading the relevant code in src/Promitor.Tests.Unit/ConfigurationKeys.cs I tried it the way I changed it to in the documentation and that worked for me. I am using the preview version for discovery (0.9.0) and for scraper (2.9.0).

ConfigurationKeys.cs: https://github.com/tomkerkhove/promitor/blob/3d176437d56e1652617812a50c1e13f9115a388b/src/Promitor.Tests.Unit/ConfigurationKeys.cs

Signed-off-by: Lino Kaiser <87489114+linokaiser@users.noreply.github.com>